% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{is.color}
\alias{is.color}
\title{is.color}
\usage{
is.color(x)
}
\arguments{
\item{x}{a vector of values}
}
\value{
a logical vector of the same length as x indicating whether each element of x is a valid color
}
\description{
is.color
}
